# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  POS-fiplex_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common_pagination.Pagination:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  POS-fiplex_internal_promotions_repository.DiscountType:
    enum:
    - percentage
    - fixed_amount
    type: string
    x-enum-varnames:
    - DiscountTypePercentage
    - DiscountTypeFixedAmount
  POS-fiplex_internal_promotions_repository.PromotionRuleType:
    enum:
    - MINIMUM_ORDER_AMOUNT
    - REQUIRED_PRODUCT
    - REQUIRED_CATEGORY
    - ALLOWED_PAYMENT_METHOD
    - ALLOWED_ORDER_TYPE
    type: string
    x-enum-varnames:
    - PromotionRuleTypeMINIMUMORDERAMOUNT
    - PromotionRuleTypeREQUIREDPRODUCT
    - PromotionRuleTypeREQUIREDCATEGORY
    - PromotionRuleTypeALLOWEDPAYMENTMETHOD
    - PromotionRuleTypeALLOWEDORDERTYPE
  POS-fiplex_internal_promotions_repository.PromotionScope:
    enum:
    - ORDER
    - ITEM
    type: string
    x-enum-varnames:
    - PromotionScopeORDER
    - PromotionScopeITEM
  POS-fiplex_internal_promotions_repository.PromotionTargetType:
    enum:
    - PRODUCT
    - CATEGORY
    type: string
    x-enum-varnames:
    - PromotionTargetTypePRODUCT
    - PromotionTargetTypeCATEGORY
...
  internal_promotions.CreatePromotionRequest:
    properties:
      description:
        type: string
      discount_type:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.DiscountType'
        enum:
        - percentage
        - fixed_amount
      discount_value:
        type: integer
      end_date:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        type: integer
      name:
        minLength: 3
        type: string
      rules:
        items:
          $ref: '#/definitions/internal_promotions.CreatePromotionRuleRequest'
        type: array
      scope:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionScope'
        enum:
        - ORDER
        - ITEM
      start_date:
        type: string
      targets:
        items:
          $ref: '#/definitions/internal_promotions.CreatePromotionTargetRequest'
        type: array
    required:
    - discount_type
    - discount_value
    - end_date
    - name
    - scope
    - start_date
    type: object
  internal_promotions.CreatePromotionRuleRequest:
    properties:
      description:
        type: string
      rule_type:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionRuleType'
        enum:
        - MINIMUM_ORDER_AMOUNT
        - REQUIRED_PRODUCT
        - REQUIRED_CATEGORY
        - ALLOWED_PAYMENT_METHOD
        - ALLOWED_ORDER_TYPE
      rule_value:
        type: string
    required:
    - rule_type
    - rule_value
    type: object
  internal_promotions.CreatePromotionTargetRequest:
    properties:
      target_id:
        type: string
      target_type:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionTargetType'
        enum:
        - PRODUCT
        - CATEGORY
    required:
    - target_id
    - target_type
    type: object
  internal_promotions.PagedPromotionResponse:
    properties:
      pagination:
        $ref: '#/definitions/POS-fiplex_internal_common_pagination.Pagination'
      promotions:
        items:
          $ref: '#/definitions/internal_promotions.PromotionResponse'
        type: array
    type: object
  internal_promotions.PromotionResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      discount_type:
        $ref: '#/definitions/POS-fiplex_internal_promotions_repository.DiscountType'
      discount_value:
        type: integer
      end_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        type: integer
      name:
        type: string
      rules:
        items:
          $ref: '#/definitions/internal_promotions.PromotionRuleResponse'
        type: array
      scope:
        $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionScope'
      start_date:
        type: string
      targets:
        items:
          $ref: '#/definitions/internal_promotions.PromotionTargetResponse'
        type: array
      updated_at:
        type: string
    type: object
  internal_promotions.PromotionRuleResponse:
    properties:
      description:
        type: string
      id:
        type: string
      rule_type:
        $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionRuleType'
      rule_value:
        type: string
    type: object
  internal_promotions.PromotionTargetResponse:
    properties:
      id:
        type: string
      target_id:
        type: string
      target_type:
        $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionTargetType'
    type: object
  internal_promotions.UpdatePromotionRequest:
    properties:
      description:
        type: string
      discount_type:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.DiscountType'
        enum:
        - percentage
        - fixed_amount
      discount_value:
        type: integer
      end_date:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        type: integer
      name:
        minLength: 3
        type: string
      rules:
        items:
          $ref: '#/definitions/internal_promotions.CreatePromotionRuleRequest'
        type: array
      scope:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_promotions_repository.PromotionScope'
        enum:
        - ORDER
        - ITEM
      start_date:
        type: string
      targets:
        items:
          $ref: '#/definitions/internal_promotions.CreatePromotionTargetRequest'
        type: array
    required:
    - discount_type
    - discount_value
    - end_date
    - name
    - scope
    - start_date
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: POS Fiplex API
  title: POS Fiplex API
  version: "1.0"
paths:
...
  /promotions:
    get:
      consumes:
      - application/json
      description: 'Get a list of promotions with pagination and optional trash filter
        (Roles: admin, manager, cashier)'
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Show trash items
        in: query
        name: trash
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Promotions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_promotions.PagedPromotionResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: List all promotions
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
      - cashier
    post:
      consumes:
      - application/json
      description: 'Create a new promotion with rules and targets (Roles: admin, manager)'
      parameters:
      - description: Promotion details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_promotions.CreatePromotionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Promotion created successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_promotions.PromotionResponse'
              type: object
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Create a new promotion
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
  /promotions/{id}:
    delete:
      consumes:
      - application/json
      description: 'Soft delete a promotion by its ID (Roles: admin, manager)'
      parameters:
      - description: Promotion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promotion deleted successfully
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
        "400":
          description: Invalid promotion ID format
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Delete (deactivate) a promotion
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
    get:
      consumes:
      - application/json
      description: 'Retrieve details of a specific promotion by its ID (Roles: admin,
        manager, cashier)'
      parameters:
      - description: Promotion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promotion retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_promotions.PromotionResponse'
              type: object
        "400":
          description: Invalid promotion ID format
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "404":
          description: Promotion not found
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get a promotion by ID
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
      - cashier
    put:
      consumes:
      - application/json
      description: 'Update details of an existing promotion by its ID (Roles: admin,
        manager)'
      parameters:
      - description: Promotion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Promotion details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_promotions.UpdatePromotionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Promotion updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_promotions.PromotionResponse'
              type: object
        "400":
          description: Invalid project ID format or request body
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "404":
          description: Promotion not found
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update a promotion
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
  /promotions/{id}/restore:
    post:
      consumes:
      - application/json
      description: 'Restore a soft-deleted promotion by its ID (Roles: admin, manager)'
      parameters:
      - description: Promotion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promotion restored successfully
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
        "400":
          description: Invalid promotion ID format
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Restore a deleted promotion
      tags:
      - Promotions
      x-roles:
      - admin
      - manager
...
swagger: "2.0"